+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 From 2019 onwards, all notable changes to tarpaulin will be documented in this
 file.
 
+## [Unreleased]
+### Changed
+- Don't disable ASLR if it's already disabled
+
 ## [0.27.0] 2023-09-17
 ### Added
 - Added `--fail-immediately` flag to abort execution the moment the first test failure occurs